+++ docs/public/embed-components/IconChevronDown-CWwLPg5K.js
@@ -1,1 +1,0 @@
-import{bt as s}from"./shared-runtime-CbINDijk.js";import{t}from"./svg-CTKt32dy.js";var r=s('<svg><path d="m6 9 6 6 6-6"></svg>',2),a=s=>t(s,r());export{a as t};